% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapBorme_func.R
\name{lectura_borme}
\alias{lectura_borme}
\title{Scrapea el Borme en PDF y devuelve un json con con el resultado.}
\usage{
lectura_borme(url = "")
}
\arguments{
\item{url}{}
}
\value{
json
}
\description{
Scrapea el Borme en PDF, devuelve un json con con los campos reflejados por empresa en el borme y lo envía a la plataforma smart city.
}
\examples{
 lectura_borme('https://www.boe.es/borme/dias/2019/01/17/pdfs/BORME-A-2019-11-48.pdf')

}
