% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapBorme_municipio_func.R
\name{lectura_borme_municipio}
\alias{lectura_borme_municipio}
\title{Scrapea los Borme de la provincia especificada y devuelve un json con con el resultado.}
\usage{
lectura_borme_municipio(url, municipio, radio, provincia, fecha_borme)
}
\arguments{
\item{url, }{municipio, radio, provincia, fecha_borme}
}
\value{
json
}
\description{
Scrapea el Borme en PDF de la provincia especificada, en base a un minicipio de referencia, un radio de distancia en km y de una fecha específica,
 devuelve un json con con los campos reflejados por empresa en el borme y lo envía a la plataforma Smart City.
}
\examples{
 lectura_borme_municipio('https://www.boe.es/borme/dias/2019/01/17/pdfs/BORME-A-2019-11-48.pdf', 'Ermua', 30, 'Bizkaia', "2019-12-19")

}
