% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NscrapBorme_func.R
\name{N_lectura_borme}
\alias{N_lectura_borme}
\title{Scrapea los Borme de las provincias especificadas y envía el resultado a la plataforma Smart City.}
\usage{
N_lectura_borme(municipio, radio, provincias, fecha = Sys.Date())
}
\arguments{
\item{provincias, }{radio, municipio, fecha}
}
\value{
json
}
\description{
Scrapea los Borme en PDF de las provincias especificadas (en caso de introducir más de una provincia estas deben estar separadas por comas),
en base a un minicipio de referencia, un radio de distancia en km y de una fecha específica (por defecto el parámetro fecha es la fecha actual),
devuelve un json con con los campos reflejados por empresa en el borme y lo envía a la plataforma Smart City.
}
\examples{
 N_lectura_borme('Ermua', 30, 'Bizkaia, Gipuzkoa, Araba')

}
